Created dedicated Admin controllers.
Various updates to views and controllers.

+++ routes/web.php
@@ -12,46 +12,46 @@
 */
 
 // Admin routes
-Route::get('/admin/login', 'SessionsController@create');
-Route::post('/admin/login', 'SessionsController@store');
-Route::get('/admin/logout', 'SessionsController@destroy');
+Route::get('/admin/login', 'Admin\SessionsController@create');
+Route::post('/admin/login', 'Admin\SessionsController@store');
+Route::get('/admin/logout', 'Admin\SessionsController@destroy');
 
-Route::get('/admin', 'AdminController@index')->name('dashboard');
+Route::get('/admin', 'Admin\AdminController@index')->name('dashboard');
 
-Route::get('/admin/announcements', 'AnnouncementController@index');
-Route::get('/admin/announcements/create', 'AnnouncementController@create');
-Route::get('/admin/announcements/{announcement}/edit', 'AnnouncementController@edit');
-Route::post('/admin/announcements', 'AnnouncementController@store');
-Route::patch('/admin/announcements/{announcement}', 'AnnouncementController@update');
-Route::delete('admin/announcements/{id}', 'AnnouncementController@destroy');
+Route::get('/admin/announcements', 'Admin\AnnouncementController@index');
+Route::get('/admin/announcements/create', 'Admin\AnnouncementController@create');
+Route::get('/admin/announcements/{announcement}/edit', 'Admin\AnnouncementController@edit');
+Route::post('/admin/announcements', 'Admin\AnnouncementController@store');
+Route::patch('/admin/announcements/{announcement}', 'Admin\AnnouncementController@update');
+Route::delete('admin/announcements/{id}', 'Admin\AnnouncementController@destroy');
 
-Route::get('/admin/schedule', 'RaceScheduleController@index');
-Route::get('/admin/schedule/create', 'RaceScheduleController@create');
-Route::get('/admin/schedule/{id}/edit', 'RaceScheduleController@edit');
-Route::post('/admin/schedule', 'RaceScheduleController@store');
-Route::patch('/admin/schedule/{id}', 'RaceScheduleController@update');
-Route::delete('admin/schedule/{id}', 'RaceScheduleController@destroy');
+Route::get('/admin/schedule', 'Admin\RaceScheduleController@index');
+Route::get('/admin/schedule/create', 'Admin\RaceScheduleController@create');
+Route::get('/admin/schedule/{id}/edit', 'Admin\RaceScheduleController@edit');
+Route::post('/admin/schedule', 'Admin\RaceScheduleController@store');
+Route::patch('/admin/schedule/{id}', 'Admin\RaceScheduleController@update');
+Route::delete('admin/schedule/{id}', 'Admin\RaceScheduleController@destroy');
 
-Route::get('/admin/classes', 'RaceClassController@index');
-Route::get('/admin/classes/create', 'RaceClassController@create');
-Route::get('/admin/classes/{id}/edit', 'RaceClassController@edit');
-Route::post('/admin/classes', 'RaceClassController@store');
-Route::patch('/admin/classes/{id}', 'RaceClassController@update');
-Route::delete('admin/classes/{id}', 'RaceClassController@destroy');
+Route::get('/admin/classes', 'Admin\RaceClassController@index');
+Route::get('/admin/classes/create', 'Admin\RaceClassController@create');
+Route::get('/admin/classes/{id}/edit', 'Admin\RaceClassController@edit');
+Route::post('/admin/classes', 'Admin\RaceClassController@store');
+Route::patch('/admin/classes/{id}', 'Admin\RaceClassController@update');
+Route::delete('admin/classes/{id}', 'Admin\RaceClassController@destroy');
 
-Route::get('/admin/results', 'RaceResultController@index');
-Route::get('/admin/results/create', 'RaceResultController@create');
-Route::get('/admin/results/{result}/edit', 'RaceResultController@edit');
-Route::post('/admin/results', 'RaceResultController@store');
-Route::patch('/admin/results/{id}', 'RaceResultController@update');
-Route::delete('admin/results/{id}', 'RaceResultController@destroy');
+Route::get('/admin/results', 'Admin\RaceResultController@index');
+Route::get('/admin/results/create', 'Admin\RaceResultController@create');
+Route::get('/admin/results/{result}/edit', 'Admin\RaceResultController@edit');
+Route::post('/admin/results', 'Admin\RaceResultController@store');
+Route::patch('/admin/results/{id}', 'Admin\RaceResultController@update');
+Route::delete('admin/results/{id}', 'Admin\RaceResultController@destroy');
 
-Route::get('/admin/resultpositions/{result}', 'RaceResultPositionController@index');
-Route::get('/admin/resultpositions/create', 'RaceResultPositionController@create');
-Route::get('/admin/resultpositions/{result}/edit', 'RaceResultPositionController@edit');
-Route::post('/admin/resultpositions', 'RaceResultPositionController@store');
-Route::patch('/admin/resultpositions/{id}', 'RaceResultPositionController@update');
-Route::delete('admin/resultpositions/{id}', 'RaceResultPositions@destroy');
+Route::get('/admin/resultpositions/{result}', 'Admin\RaceResultPositionController@index');
+Route::get('/admin/resultpositions/create', 'Admin\RaceResultPositionController@create');
+Route::get('/admin/resultpositions/{result}/edit', 'Admin\RaceResultPositionController@edit');
+Route::post('/admin/resultpositions', 'Admin\RaceResultPositionController@store');
+Route::patch('/admin/resultpositions/{id}', 'Admin\RaceResultPositionController@update');
+Route::delete('admin/resultpositions/{id}', 'Admin\RaceResultPositions@destroy');
 
 Route::get('/admin/racers/{result}', 'RacerController@index');
 Route::get('/admin/racers/create', 'RacerController@create');
@@ -67,15 +67,24 @@ Route::get('/', 'HomeController@index');
 //    return view('home');
 //});
 
-Route::get('/schedule', 'RaceScheduleController@view');
+Route::get('/schedule', 'RaceScheduleController@index');
 
-Route::get('/results', 'RaceResultController@view');
+Route::get('/results', 'RaceResultController@index');
+Route::get('/results/{event}', 'RaceResultController@show');
+
 Route::get('/point-standings', function () {
     return view('point_standings.index');
 });
 
-Route::get('/announcements', 'AnnouncementController@view');
+Route::get('/announcements', 'AnnouncementController@index');
 Route::get('/announcements/{announcement}', 'AnnouncementController@show');
+
+Route::get('/directions', function() {
+   return view('directions.index');
+});
+Route::get('/about', function() {
+    return view('about.index');
+});
 
 // Eloquent model => Post - usually a noun (doesn't have to be)
 // controller => PostsController
